gremlin-javascript: Tune Writer/Reader abstraction border
